# api/swagger/openapi.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Comprehensive API for monitoring and managing NATS JetStream instances.",
    "title": "NATS Horizon API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "NATS Horizon Project",
      "url": "http://github.com/nats-horizon"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    },
    "version": "1.0"
  },
  "basePath": "/api",
  "paths": {
...
    "/history/report": {
      "get": {
        "description": "Returns a summary report of all streams with latest metric values",
        "produces": [
          "application/json"
        ],
        "tags": [
          "history"
        ],
        "summary": "Get history report",
        "parameters": [
          {
            "type": "string",
            "default": "24h",
            "description": "Time window",
            "name": "duration",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "history report",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/history/streams/{name}": {
      "get": {
        "description": "Returns historical metric data points for a stream",
        "produces": [
          "application/json"
        ],
        "tags": [
          "history"
        ],
        "summary": "Get stream history",
        "parameters": [
          {
            "type": "string",
            "description": "Stream name",
            "name": "name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "default": "messages",
            "description": "Metric type (messages, bytes)",
            "name": "type",
            "in": "query"
          },
          {
            "type": "string",
            "default": "24h",
            "description": "Time window (1h, 6h, 24h, 7d)",
            "name": "duration",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "stream history",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/history/streams/{name}/analysis": {
      "get": {
        "description": "Returns statistical analysis (min/max/avg/trend) of a stream metric",
        "produces": [
          "application/json"
        ],
        "tags": [
          "history"
        ],
        "summary": "Get stream metric analysis",
        "parameters": [
          {
            "type": "string",
            "description": "Stream name",
            "name": "name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "default": "messages",
            "description": "Metric type (messages, bytes)",
            "name": "type",
            "in": "query"
          },
          {
            "type": "string",
            "default": "24h",
            "description": "Time window (1h, 6h, 24h, 7d)",
            "name": "duration",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "metric analysis",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
  },
...
}